# server/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "FastFood API",
    "description": "Swagger Docs delle API di FastFood"
  },
  "servers": [
    {
      "url": "http://localhost:30000/"
    }
  ],
  "paths": {
...
    "/menu": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/menu/add": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "piattoComuneId": {
                    "example": "any"
                  },
                  "piattoPersonalizzato": {
                    "example": "any"
                  },
                  "strMeal": {
                    "example": "any"
                  },
                  "ingredients": {
                    "example": "any"
                  },
                  "strCategory": {
                    "example": "any"
                  },
                  "strMealThumb": {
                    "example": "any"
                  },
                  "prezzo": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/menu/update/{menuItemId}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "menuItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prezzo": {
                    "example": "any"
                  },
                  "strMeal": {
                    "example": "any"
                  },
                  "strCategory": {
                    "example": "any"
                  },
                  "ingredients": {
                    "example": "any"
                  },
                  "strMealThumb": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/menu/remove/{menuItemId}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "menuItemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/menu/restaurant/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}